Implementado el patron Factory Method
