--- src/coloniaDeRobots/CofreAlmacenamiento.java
@@ -1,7 +1,0 @@
-package coloniaDeRobots;
-
-public class CofreAlmacenamiento extends Cofre{
-	CofreAlmacenamiento(int coordX,int coordY){
-		super(coordX,coordY);
-	}
-}
