+++ src/coloniaDeRobots/CofreAlmacenamientoFactory.java
@@ -1,0 +1,10 @@
+package coloniaDeRobots;
+import coloniaDeRobots.Cofres.Cofre;
+import coloniaDeRobots.Cofres.CofreAlmacenamiento;
+
+public class CofreAlmacenamientoFactory extends CofreFactory{
+	@Override
+	public Cofre crearCofre(int coordX,int coordY) {
+		return new CofreAlmacenamiento(coordX,coordY);
+	}
+}
