+++ src/coloniaDeRobots/CofreFactory.java
@@ -1,5 +1,8 @@
 package coloniaDeRobots;
 
-public abstract class CofreFactory {
+import coloniaDeRobots.Cofres.Cofre;
 
+public abstract class CofreFactory {
+	abstract Cofre crearCofre(int coordX,int coordY);
 }
+
