--- src/coloniaDeRobots/CofreIntermedio.java
@@ -1,8 +1,0 @@
-package coloniaDeRobots;
-
-public class CofreIntermedio extends Cofre{
-	CofreIntermedio(int coordX,int coordY){
-		super(coordX,coordY);
-	}
-}
-
