+++ src/coloniaDeRobots/CofreIntermedioFactory.java
@@ -1,5 +1,10 @@
 package coloniaDeRobots;
+import coloniaDeRobots.Cofres.Cofre;
+import coloniaDeRobots.Cofres.CofreIntermedio;
 
-public class CofreIntermedioFactory {
-
+public class CofreIntermedioFactory extends CofreFactory{
+	@Override
+	public Cofre crearCofre(int coordX,int coordY) {
+		return new CofreIntermedio(coordX,coordY);
+	}	
 }
