--- src/coloniaDeRobots/CofreProvisionActiva.java
@@ -1,7 +1,0 @@
-package coloniaDeRobots;
-
-public class CofreProvisionActiva extends Cofre {
-	CofreProvisionActiva(int coordX,int coordY){
-		super(coordX,coordY);
-	}
-}
