--- src/coloniaDeRobots/CofreProvisionPasiva.java
@@ -1,7 +1,0 @@
-package coloniaDeRobots;
-
-public class CofreProvisionPasiva extends Cofre{
-	CofreProvisionPasiva(int coordX,int coordY){
-		super(coordX,coordY);
-	}
-}
