+++ src/coloniaDeRobots/CofreProvisionPasivaFactory.java
@@ -1,0 +1,10 @@
+package coloniaDeRobots;
+import coloniaDeRobots.Cofres.Cofre;
+import coloniaDeRobots.Cofres.CofreProvisionPasiva;
+
+public class CofreProvisionPasivaFactory extends CofreFactory{
+	@Override
+	public Cofre crearCofre(int coordX,int coordY) {
+		return new CofreProvisionPasiva(coordX,coordY);
+	}	
+}
