--- src/coloniaDeRobots/CofreSolicitud.java
@@ -1,7 +1,0 @@
-package coloniaDeRobots;
-
-public class CofreSolicitud extends Cofre{
-	CofreSolicitud(int coordX,int coordY){
-		super(coordX,coordY);
-	}
-}
