+++ src/coloniaDeRobots/Cofres/CofreAlmacenamiento.java
@@ -1,0 +1,7 @@
+package coloniaDeRobots.Cofres;
+
+public class CofreAlmacenamiento extends Cofre{
+	public CofreAlmacenamiento(int coordX,int coordY){
+		super(coordX,coordY);
+	}
+}
