+++ src/coloniaDeRobots/Cofres/CofreIntermedio.java
@@ -1,0 +1,8 @@
+package coloniaDeRobots.Cofres;
+
+public class CofreIntermedio extends Cofre{
+	public CofreIntermedio(int coordX,int coordY){
+		super(coordX,coordY);
+	}
+}
+
