+++ src/coloniaDeRobots/Cofres/CofreProvisionActiva.java
@@ -1,0 +1,7 @@
+package coloniaDeRobots.Cofres;
+
+public class CofreProvisionActiva extends Cofre {
+	public CofreProvisionActiva(int coordX,int coordY){
+		super(coordX,coordY);
+	}
+}
