+++ src/coloniaDeRobots/Cofres/CofreProvisionPasiva.java
@@ -1,0 +1,7 @@
+package coloniaDeRobots.Cofres;
+
+public class CofreProvisionPasiva extends Cofre{
+	public CofreProvisionPasiva(int coordX,int coordY){
+		super(coordX,coordY);
+	}
+}
