+++ src/coloniaDeRobots/Cofres/CofreSolicitud.java
@@ -1,0 +1,7 @@
+package coloniaDeRobots.Cofres;
+
+public class CofreSolicitud extends Cofre{
+	public CofreSolicitud(int coordX,int coordY){
+		super(coordX,coordY);
+	}
+}
